Redo migrations from scratch to change id field type

+++ db/migrations/20210515120057_link_squares_and_themes.ts
@@ -3,9 +3,9 @@ import { Knex } from "knex";
 
 export async function up(knex: Knex): Promise<void> {
     return knex.schema.createTable("squares_themes", (table) => {
-        table.bigIncrements();
-        table.bigInteger("square_id").references("squares.id").unsigned().index().onDelete("CASCADE");
-        table.bigInteger("theme_id").references("themes.id").unsigned().index().onDelete("CASCADE");
+        table.increments();
+        table.integer("square_id").references("squares.id").unsigned().index().onDelete("CASCADE");
+        table.integer("theme_id").references("themes.id").unsigned().index().onDelete("CASCADE");
     });
 }
